+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * 【core   】     增加CharSequenceUtil.subPreGbk重载（issue#I4JO2E@Gitee）
 * 【core   】     ReflectUtil.getMethod排除桥接方法（pr#1965@Github）
 * 【http   】     completeFileNameFromHeader在使用path为路径时，自动解码（issue#I4K0FS@Gitee）
+* 【core   】     CopyOptions增加override配置（issue#I4JQ1N@Gitee）
 * 
 ### 🐞Bug修复
 * 【core   】     修复FileResource构造fileName参数无效问题（issue#1942@Github）